Create W2-13
Write a program that reads in the principal amount in a saving account, its interest rate, and the number
of years since the account was opened. Your program then should calculate and print the total amount
(i.e. the principal amount plus the interest) in the saving account after the time period specified. You must
decide what data types are appropriate for your variables. Use simple interest (not compound interest)