Create W2-14
 Write a complete C++ program that reads two integer values in the range [1, 30] representing calendar
days from the user and then prints the number of days between the two calendar days. For example if you
enter 5 for the first variable and 27 for the second variable, then your program must print "There are 22
days between day 5 and day 27". Of course it is ok if your program prints the days as a negative number
such as "There are -22 days between day 5 and day 27". The reason why you may get negative output is
that you don't know if you will be subtracting the smaller day from the larger or vice versa; after all you
don't know which variable will get the larger day and which one gets the lesser day until you run your
program. Assume the input values for the days are valid; i.e. in the range [1, 30].