Create W2-15
 Write a complete C++ program that reads two integer values in the range [1, 12] representing calendar
months from the user and then prints the number of days between the two calendar months. Assume a
month has 30 days. For example if you enter 5 for the first variable and 1 for the second variable, then
your program must print "There are 120 days between month 5 and month 1". Of course it is ok if your
program prints the days as a negative number such as "There are -120 days between month 5 and month
1". Assume the input values for the months are valid; i.e. in the range [1, 12].