Create W2-16
Write a complete C++ program that reads two integer values representing calendar years from the user
and then prints the number of days between the two calendar years. Assume a year has 360 days. For
example if you enter 1987 for the first variable and 2015 for the second variable, then your program must
print "There are 10080 days between year 1987 and year 2015". Of course it is ok if your program prints
the days as a negative number such as "There are -10080 days between year 1987 and year 2015". Assume
the input values for the years are valid; i.e. they are non-negative integers.