Create W2-17
Write a C++ program that declares six variables named y1, m1, d1, y2, m2, and d2 all as integer data types.
Now read the birth date of a child 1 in y1, m1 and d1 variables where y1, m1 and d1 represent the year,
month and day of birth date of child 1. Then read the birth date of child 2 in y2, m2, and d2 variables.
Finally print the number of days between the birth dates of the two children. It is ok if your program prints
the number of days as negative or positive. Assume the input values for days are in the range [1, 30], the
input values for months are in the range [1, 12], and the input values for the years are some positive
numbers. This means a month has 30 days and a year has 12 months ( = 360 days).