Create W2-18
Repeat Q17 but this time your program must print the number of years, number of months and numbers
of days between the birth dates of the two children. Assume a year has 12 months ( = 360 days) and a
month has 30 days. Your output must contain only positive numbers and zero OR only negative numbers
and zero. NO MIX OF NEGATIVE AND POSITIVE OUTPUTS IS ALLOWED.