Create W3-12(Pizza to buy)
The large “economy” size of an item is not always a better buy than
the smaller size. This is particularly true when buying pizzas. Pizza sizes are given as the
diameter of the pizza in inches. However, the quantity of pizza is determined by the area of
the pizza, and the area is not proportional to the diameter. Most people cannot easily
estimate the difference in area between a 10-inch pizza and a 12-inch pizza and so cannot
easily determine which size is the best buy—that is, which size has the lowest price per
square inch