Create W3-15(BUG)
Write a C++ program that generates a random upper case English alphabet character and
that prints the upper case English alphabet character that is ten letters away from the
character generated. For example,
If the random character is 'A' then your program must print 'K'
If the random character is 'B' then your program must print 'L'
If the random character is 'C' then your program must print 'M'
If the random character is 'Q' then your program must print 'A'
If the random character is 'V' then your program must print 'F'
If the random character is 'Z' then your program must print 'J'
Observe that if the character that is ten letters away from the random character is outside
of the upper case English letters; then we must go back to 'A'.