Create W4-7
Write a program that reads in an integer n, asserts n is positive (that is the program uses a
while loop structure to repeatedly ask the user for n until the user enters a positive n), and
then reads n integers from the user using a while loop. Your program should then output
the sum of all even numbers, the average of all even numbers, the sum of all odd numbers,
the average of all odd numbers, the sum of all the numbers, and the average of all numbers
entered. Every negative entry should be skipped and not used in any of the calculations