Create W4-9
Write a C++ program that reads a positive integer user input n and then that reads n user
input integers and finally prints the maximum of all the n user input integers. Assume the
user input value for n is positive. For example if the value of n is 5 and the numbers are -9, -
3, -7, -2, -6 then your program must print the maximum value is -2