Create W4-11
Write a C++ program that reads a positive integer user input n and then that reads n user
input integers and finally prints the maximum in absolute value of all the n user input
integers. Assume the user input value for n is positive. For example if n is 4 and the user
input are 2, -3, 4, -6 then your program must print The maximum in absolute value is -6.
For example if the value of n is 5 and the numbers are 9, -3, -7, -25, -6 then your program
must print the maximum in absolute value is -25