Create W4-12
Write a program that reads a positive integer number n and then prints the double data
type value y given by:
𝑦 = 1 − 1/2 + 1/3 − 1/4 + 1/5 − 1/6 + ⋯ ± 1/𝑛
Note that the last term may have a plus sign or a minus sign depending on the value of n.
Use a while loop. Assume that the user input value for n is positive integer. Also note that
the divisions must be performed in a float/double domain to avoid truncation.