Create W4-13
Write a C++ program that reads a positive integer n and prints "The number is prime" if n is
a prime number and prints "The number is not prime" if n is not a prime number. Assume
the user input is an integer greater than 2. Hint:- use a while loop to count how many
divisors n has between 2 and n-1. If it has zero divisors, it is a prime number; otherwise it is
not a prime number