Create W4-14
Write a C++ program that reads a positive integer n and prints "The number is composite"
if n is a composite number and prints "The number is not composite" if n is not a composite
number. Hint:- Note that a number is composite if it is not prime; and vice versa. Use your
ideas in Question #13 above