Create W5-4
Write a complete C++ program that reads two integers a and b and then prints all the
integers between a and b (or vice versa) exclusive; that is without including a and b using a
do-while loop. Remember if either the value of a is equal to b or that a and b are
consecutive integers then your program should not print any integer because under such
circumstances there is no any integer between a and b. Under such circumstance, your
program should instead print the message "There is no any integer"