Create W6-18
Area of Triangle:- Consider the area of a triangle whose three vertices on the plane are given as points
P1(x1,y1), P2(x2,y2),and P3(x3,y3). Write a C++ function named areaOfTriangle that takes six
arguments x1, y1, x2, y2, x3, and y3 and returns the area of the triangle. Assume no two points are
collinear. Hint:- First calculate the length of each of the sides of the triangle using the Euclidean Distance
formula. For this you should make use of your function vectorLength