Create W6-19
Simple Interest Calculator:- Write a C++ function named accruedTotal that that takes two float
arguments: a principal amount and time period in years and returns the accrued total amount. The
accrued total amount is calculated as
totalAmount = principalAmount * (1.0 + rateOfInterest * timePeriod)
For the rate of interest use the following table:
Principal Amount Rate of Interest
[0.00, 1000.00) 2.5%
[1000.0, 10,000.00) 2.0%
[10,000.00, 100,000.00) 1.5%
≥ 100,000.00 1.0%
The main program is given below in order to test your function.