Create W6-20
Polynomial Function:- Given the cubic polynomial function
f(x) = ax3 + bx2 + cx + d
Write a C++ function named f that takes five arguments a, b, c, d, and x where a, b, c, and d are the
coefficients of the cubic polynomial and x is a point whose f(x) we would like to compute, and returns the
value of f(x). In order to make it a general C++ function that works for any cubic order polynomial; this
function must take all the coefficients a, b, c, and d as arguments. In addition, the function has to take the
value of x as an argument. The function returns f(x)