Create W7-9
Write a C++ function named countElementsFromIndex that takes four arguments: an array of integers, its
size, an integer value x, and an integer index. Your function must search the integer value x in the array
starting from the index argument upto the last element of the array and return the number of elements in
the array that are equal to the integer argument. For example give the following code
const int size = 11;
int A[]= {5, 7, 8, 0, 5, 4, 1, 5, 6, 9, 5};
int searchValue = 5;
int startIndex = 2;
Then the function call countElementsFromIndex(A, size, searchValue, startIndex); must
return 3 because there are 3 elements equal to 5 starting from index 2. Of course in total there are 4
elements equal to 5 but because the search starts at the index given by the index argument, then the
function will find only 3 elements equal to 5