Create W7-10
Write a C++ function named countCommonElements that takes four arguments: an array of integers, its
size, a second array of integers, and its size. Your function must return the number of elements of the first
array that are also found in the second array. Assume each of the arrays contains distinct (i.e. different)
elements.